+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn application:application --log-file=-
+web: gunicorn --worker-class eventlet -w 1 application:application --log-file=-
