+++ application.py
@@ -1,7 +1,6 @@
 from service_application_package import create_app
-
+from service_application_package.chat.chatroute import socketio
 application = create_app()
 
 if __name__ == '__main__':
-	
-	application.run()
+    socketio.run(application)
