+++ requirements.txt
@@ -6,6 +6,7 @@ Flask==1.0.2
 Flask-Bcrypt==0.7.1
 Flask-Login==0.4.1
 Flask-Mail==0.9.1
+Flask-Migrate==2.4.0
 Flask-SocketIO==3.3.2
 Flask-SQLAlchemy==2.3.2
 Flask-WTF==0.14.2
@@ -14,6 +15,9 @@ greenlet==0.4.15
 gunicorn==19.9.0
 itsdangerous==1.1.0
 Jinja2==2.10
+jmespath==0.9.4
+jsonschema==2.6.0
+Mako==1.0.8
 MarkupSafe==1.1.1
 Pillow==5.4.1
 portalocker==1.4.0
@@ -23,9 +27,15 @@ python-engineio==3.5.0
 python-socketio==3.1.2
 PyYAML==5.1
 redis==3.2.1
-six==1.12.0
+requests==2.20.1
+rsa==3.4.2
+s3transfer==0.2.0
+semantic-version==2.5.0
+six==1.11.0
 SQLAlchemy==1.3.1
 urllib3==1.24.1
+wcwidth==0.1.7
+websocket-client==0.56.0
 Werkzeug==0.15.1
 WTForms==2.2.1
 WTForms-SQLAlchemy==0.1
